+++ src/main/java/com/iexec/sms/admin/AdminService.java
@@ -41,6 +41,10 @@ public class AdminService {
 
     // Used to print formatted date in log
     private final DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+
+    public static final String ERR_BACKUP_FILE_OUTSIDE_STORAGE = "Backup file is outside of storage file system";
+    public static final String ERR_BACKUP_FILE_NOT_EXIST = "Backup file does not exist";
+    public static final String ERR_FILE_ALREADY_EXIST = "A file already exists at the destination";
     private final String datasourceUrl;
     private final String datasourceUsername;
     private final String datasourcePassword;
@@ -156,9 +160,9 @@ public class AdminService {
         try {
             final String backupFileLocation = checkBackupFileLocation(
                     storageLocation + File.separator + backupFileName,
-                    "Backup file is outside of storage file system");
+                    ERR_BACKUP_FILE_OUTSIDE_STORAGE);
             if (!Path.of(backupFileLocation).toFile().exists()) {
-                throw new FileSystemNotFoundException("Backup file does not exist");
+                throw new FileSystemNotFoundException(ERR_BACKUP_FILE_NOT_EXIST);
             }
             final long size = backupFileLocation.length();
             log.info("Starting the restore process [backupFileLocation:{}]", backupFileLocation);
@@ -193,10 +197,10 @@ public class AdminService {
             }
             final String backupFileLocation = checkBackupFileLocation(
                     storageLocation + File.separator + backupFileName,
-                    "Backup file is outside of storage file system");
+                    ERR_BACKUP_FILE_OUTSIDE_STORAGE);
             final Path backupFileLocationPath = Path.of(backupFileLocation);
             if (!backupFileLocationPath.toFile().exists()) {
-                throw new FileSystemNotFoundException("Backup file does not exist");
+                throw new FileSystemNotFoundException(ERR_BACKUP_FILE_NOT_EXIST);
             }
             log.info("Starting the delete process [backupFileLocation:{}]", backupFileLocation);
             final long start = System.currentTimeMillis();
@@ -207,6 +211,52 @@ public class AdminService {
             return true;
         } catch (IOException e) {
             log.error("An error occurred while deleting backup", e);
+        }
+        return false;
+    }
+
+    /**
+     * Copy a backup from a location to another with the possibility of renaming the file
+     *
+     * @param sourceStorageLocation      The location of the source backup file.
+     * @param sourceBackupFileName       The name of the source backup file.
+     * @param destinationStorageLocation The location of destination the backup file.
+     * @param destinationBackupFileName  The name of the destination backup file.
+     * @return {@code true} if the copy was successful, {@code false} if any error occurs.
+     */
+    boolean copyBackupFileFromStorageToStorage(String sourceStorageLocation, String sourceBackupFileName, String destinationStorageLocation, String destinationBackupFileName) {
+        try {
+            // Check that we want to copy an authorised file
+            final Path sourceBackupFileLocation = Path.of(checkBackupFileLocation(
+                    sourceStorageLocation + File.separator + sourceBackupFileName,
+                    ERR_BACKUP_FILE_OUTSIDE_STORAGE));
+            // File must exist
+            if (!sourceBackupFileLocation.toFile().exists()) {
+                throw new FileSystemNotFoundException(ERR_BACKUP_FILE_NOT_EXIST);
+            }
+            // No renaming, keep the original name
+            if (StringUtils.isBlank(destinationBackupFileName)) {
+                destinationBackupFileName = sourceBackupFileName;
+            }
+            // Check that we want to copy into authorized location
+            final Path destinationBackupFileLocation = Path.of(checkBackupFileLocation(
+                    destinationStorageLocation + File.separator + destinationBackupFileName,
+                    ERR_BACKUP_FILE_OUTSIDE_STORAGE));
+
+            // Check that we are not trying to overwrite a file
+            if (destinationBackupFileLocation.toFile().exists()) {
+                throw new IOException(ERR_FILE_ALREADY_EXIST);
+            }
+
+            log.info("Starting the copy process [sourceBackupFileLocation:{}, destinationBackupFileLocation:{}]", sourceBackupFileLocation, destinationBackupFileLocation);
+            final long start = System.currentTimeMillis();
+            Files.copy(sourceBackupFileLocation, destinationBackupFileLocation, StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.COPY_ATTRIBUTES);
+            final long stop = System.currentTimeMillis();
+            log.info("Backup has been copied [sourceBackupFileLocation:{}, destinationBackupFileLocation:{}, timestamp:{}, duration:{} ms]",
+                    sourceBackupFileLocation, destinationBackupFileLocation, dateFormat.format(start), stop - start);
+            return true;
+        } catch (IOException e) {
+            log.error("An error occurred while copying backup", e);
         }
         return false;
     }