+++ src/test/java/com/iexec/sms/admin/AdminControllerTests.java
@@ -351,6 +351,80 @@ class AdminControllerTests {
     }
     // endregion
 
+    // region replicate-backup
+    @Test
+    void testCopy(@TempDir Path tempDir) {
+        final String sourceStorageID = convertToHex(tempDir.toString());
+
+        ReflectionTestUtils.setField(adminController, "adminStorageLocation", tempDir.toString());
+        when(adminService.copyBackupFileFromStorageToStorage(tempDir.toString(), FILE_NAME, tempDir.toString(), "backup2.sql")).thenReturn(true);
+        assertEquals(HttpStatus.OK, adminController.copyBackup(sourceStorageID, sourceStorageID, FILE_NAME, "backup2.sql").getStatusCode());
+    }
+
+    @Test
+    void testInterruptedThreadOnCopy() throws InterruptedException {
+        ReflectionTestUtils.setField(adminController, "rLock", rLock);
+        when(rLock.tryLock(100, TimeUnit.MILLISECONDS)).thenThrow(InterruptedException.class);
+        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, adminController.copyBackup(STORAGE_PATH, STORAGE_PATH, FILE_NAME, FILE_NAME).getStatusCode());
+    }
+
+    @Test
+    void testInternalServerErrorOnCopy(@TempDir Path tempDir) {
+        final String sourceStorageID = convertToHex(tempDir.toString());
+        when(adminService.copyBackupFileFromStorageToStorage(tempDir.toString(), FILE_NAME, tempDir.toString(), "backup2.sql")).thenThrow(RuntimeException.class);
+        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, adminController.copyBackup(sourceStorageID, sourceStorageID, FILE_NAME, "backup2.sql").getStatusCode());
+    }
+
+    @Test
+    void testTooManyRequestOnCopy(@TempDir Path tempDir) throws InterruptedException {
+
+        AdminController adminControllerWithLongAction = new AdminController(new AdminService("", "", "", "") {
+            @Override
+            public boolean copyBackupFileFromStorageToStorage(String sourceStorageLocation, String sourceBackupFileName, String destinationStorageLocation, String destinationBackupFileName) {
+                try {
+                    log.info("Long copyBackupFileFromStorageToStorage action is running ...");
+                    Thread.sleep(2000);
+                } catch (InterruptedException e) {
+                    Thread.currentThread().interrupt();
+                }
+                return true;
+            }
+        }, "");
+
+        final List<ResponseEntity<Void>> responses = Collections.synchronizedList(new ArrayList<>(3));
+        final String sourceStorageID = convertToHex(tempDir.toString());
+
+        Thread firstThread = new Thread(() -> responses.add(adminControllerWithLongAction.copyBackup(sourceStorageID, sourceStorageID, FILE_NAME, "backup2.sql")));
+        Thread secondThread = new Thread(() -> responses.add(adminControllerWithLongAction.copyBackup(sourceStorageID, sourceStorageID, FILE_NAME, "backup2.sql")));
+
+        firstThread.start();
+        secondThread.start();
+        responses.add(adminControllerWithLongAction.copyBackup(sourceStorageID, sourceStorageID, FILE_NAME, "backup2.sql"));
+
+        secondThread.join();
+        firstThread.join();
+
+
+        long code200 = responses.stream().filter(element -> element.getStatusCode() == HttpStatus.OK).count();
+        long code429 = responses.stream().filter(element -> element.getStatusCode() == HttpStatus.TOO_MANY_REQUESTS).count();
+
+        assertEquals(1, code200);
+        assertEquals(2, code429);
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideBadRequestParameters")
+    void testBadRequestOnCopyOnCommonsAdminControls(String storageID, String fileName) {
+        assertEquals(HttpStatus.BAD_REQUEST, adminController.copyBackup(storageID, "NOT_CONCERNED", fileName, "NOT_CONCERNED").getStatusCode());
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideBadRequestParametersForCopy")
+    void testBadRequestOnCopyOnSpecificsCopyControls(String sourceStorageID, String sourceFileName, String destinationStorageID) {
+        assertEquals(HttpStatus.BAD_REQUEST, adminController.copyBackup(sourceStorageID, destinationStorageID, sourceFileName, "NOT_CONCERNED").getStatusCode());
+    }
+    // endregion
+
     private static Stream<Arguments> provideBadRequestParameters() {
         return Stream.of(
                 Arguments.of(null, null),
@@ -358,6 +432,14 @@ class AdminControllerTests {
                 Arguments.of(" ", " "),
                 Arguments.of(STORAGE_PATH, " "),
                 Arguments.of(" ", FILE_NAME)
+        );
+    }
+
+    private static Stream<Arguments> provideBadRequestParametersForCopy() {
+        return Stream.of(
+                Arguments.of(STORAGE_PATH, FILE_NAME, ""),
+                Arguments.of(STORAGE_PATH, FILE_NAME, " "),
+                Arguments.of(STORAGE_PATH, FILE_NAME, null)
         );
     }
 }