+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Add the ability to trigger a database restore via a dedicated endpoint. (#212)
 - Add the ability to trigger a delete via a dedicated endpoint. (#213)
 - Add the ability to trigger a backup replication via a dedicated endpoint. (#214)
+- Add the ability to trigger a backup copy via a dedicated endpoint. (#217)
 
 ### Dependency Upgrades